version 1.4.8
Fix Kernel 32 Bits only can select 32 Bits. If you select more returns core dump.
Now returns message if selected region is bigger than 32 Bits on 32 Bits kernel.

+++ hexedit.h
@@ -200,7 +200,7 @@ int not(int b);
 char *mymemmem(char *a, size_t sizea, char *b, size_t sizeb);
 char *mymemrmem(char *a, size_t sizea, char *b, size_t sizeb);
 int is_file(char *name);
-int hexStringToBinString(char *p, size_t *l);
+int hexStringToBinString(char *p, INT *l);
 
 /*******************************************************************************/
 /* Functions provided for OSs that don't have them */
